Refactor files a bit
