+++ packages/cloud_firestore_odm/cloud_firestore_odm_generator/lib/src/templates/collection_reference.dart
@@ -1,3 +1,4 @@
+import '../collection_data.dart';
 import '../collection_generator.dart';
 
 class CollectionReferenceTemplate {
